moved search component, added in nav bar
