+++ app/components/Navigation.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+export const Navigation = () => {
+    return (
+      <>
+	<ul>
+	  <li></li>
+	  <li></li>
+	  <li></li>
+	  <li></li>
+	  <li></li>
+	</ul>
+      </>
+    )
+}
