+++ app/components/tooltip.jsx
@@ -1,11 +1,12 @@
 import React from 'react'
 import { Tooltip } from 'antd'
+import CustomSearch from './search'
 import { Html } from '@react-three/drei'
 
 export const CustomTooltip = () => {
   return (
     <Html center>
-      <div style={{display: 'flex', marginTop: '20em', gap: '50px' }}>
+      <div style={{display: 'flex', gap: '50px' }}>
 	<Tooltip title="Type 'cd about' to see who I am, what I have worked on, and what I enjoy!">
 	    <span> Go to my About Me page.</span>
 	</Tooltip>
